add nbformat.sign.NotebookNotary
